vinscant: refactor code
