+++ vinscant/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod card_request;
 pub mod ping_pong;
 pub mod wifi;
